refactor municipality button since it is reused
